change double to int

+++ services/product/src/main/java/org/microservices/product/entities/Product.java
@@ -20,7 +20,7 @@ public class Product {
     private Integer id;
     private String name;
     private String description;
-    private double availableQuantity;
+    private int availableQuantity;
     private BigDecimal price;
 
     @ManyToOne
